Fix "constant overflows int" compile error on 32-bit systems
When passing the constant `jks.MagicNumber` to a variadic function
through interface{}, the compiler selects type `int`. On a 32-bit
machine, the constant `jks.MagicNumber` does not fit into a signed
32-bit integer.

Give the constant an explicit type to avoid this problem. Not only does
this resolve the specific issue, it also guards any other potential
misuses of the constant.

With thanks to @Xianic.